Blosc_widget_upgraded
